first chain commit

+++ src/main/java/com/epl/protobuf/ShapefileChunkedReader.java
@@ -162,7 +162,7 @@ public class ShapefileChunkedReader extends GeometryCursor {
     }
 
     @Override
-    public long getGeometryID() {
+    public int getGeometryID() {
         return currentRecordNumber;
     }
